backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
google-genai
chromadb
pydantic
